Move models to separate folder
